Implement a cross-validated grid search model
The assumption of this training is to use a DecisionTreeRegressor.

Integer values of max_depth will be varied from 1 to 10, inclusive,
during the grid search for the best parameters to use in the classifer.

The scoring function has already been implemented in performance_metric.
Thus, that function is made into a scorer for use during the GridSearch.

Here also, cv is passed an interable, 'cv_sets', which contains the
indexes of the data points to be used in training.

The grid search is conducted and the best estimator is returned.