Remove Bessel's Correction from STDev calculation
Previously, Bessel's Correction was used to calculate the sample
standard deviation.  This is tenuous at best for standard deviation,
since the estimator it yields is still biased, though less biased then
not using Bessel's Correction.

The unbiaser for Standard Deviation is actually quite complicated, and
is described at
https://en.wikipedia.org/wiki/Unbiased_estimation_of_standard_deviation#Results_for_the_normal_distribution
This of course presumes a normal distribution, which it's not clear
applies to home prices.  Maybe it does, but we certainly haven't checked
it for this dataset.

Further, for datasets of n >> 75, it's "generally acceptable" to use the
uncorrected sample standard deviation, since (1) the bias asymptotically
approaches 0 as 1/n, and (2) the uncorrected sample standard deviation
estimator has a smaller mean squared error than does the corrected
sample standard deviation estimator. See
https://en.wikipedia.org/wiki/Standard_deviation#Uncorrected_sample_standard_deviation

Hence, change ddof=1 to ddof=0, and concisely explain the logic for
doing so.